fix(review): apply critical and high-severity review findings
Race condition:
- CreateOrderUseCase no longer holds shared OrderCreated event
- event.NewOrderCreated() instantiated per Execute() call (thread-safe)

Error handling:
- Handlers return generic messages, log details server-side (CWE-209)
- http.MaxBytesReader(1MB) on REST Create handler (CWE-20)
- webserver.Start() uses log.Fatalf instead of silently discarding error
- main.go uses log.Fatalf instead of panic (goroutine-safe)
- db.Ping() validates MySQL connection after sql.Open()
- eventDispatcher.Register error checked in main.go

gRPC precision:
- proto fields changed from float to double (float64) — no more truncation
- service/order_service.go: removed float32 casts

Dependencies:
- Migrated streadway/amqp (archived) to rabbitmq/amqp091-go v1.10.0

+++ internal/infra/grpc/service/order_service.go
@@ -26,8 +26,8 @@ func NewOrderGrpcService(
 func (s *OrderGrpcService) CreateOrder(ctx context.Context, in *pb.CreateOrderRequest) (*pb.CreateOrderResponse, error) {
 	dto := usecase.OrderInputDTO{
 		ID:    in.Id,
-		Price: float64(in.Price),
-		Tax:   float64(in.Tax),
+		Price: in.Price,
+		Tax:   in.Tax,
 	}
 	output, err := s.CreateOrderUseCase.Execute(dto)
 	if err != nil {
@@ -35,9 +35,9 @@ func (s *OrderGrpcService) CreateOrder(ctx context.Context, in *pb.CreateOrderRe
 	}
 	return &pb.CreateOrderResponse{
 		Id:         output.ID,
-		Price:      float32(output.Price),
-		Tax:        float32(output.Tax),
-		FinalPrice: float32(output.FinalPrice),
+		Price:      output.Price,
+		Tax:        output.Tax,
+		FinalPrice: output.FinalPrice,
 	}, nil
 }
 
@@ -50,9 +50,9 @@ func (s *OrderGrpcService) ListOrders(ctx context.Context, in *pb.ListOrdersRequ
 	for _, o := range output {
 		orders = append(orders, &pb.CreateOrderResponse{
 			Id:         o.ID,
-			Price:      float32(o.Price),
-			Tax:        float32(o.Tax),
-			FinalPrice: float32(o.FinalPrice),
+			Price:      o.Price,
+			Tax:        o.Tax,
+			FinalPrice: o.FinalPrice,
 		})
 	}
 	return &pb.ListOrdersResponse{Orders: orders}, nil
